Refactored code to cleanup game.js
