+++ public/js/config.js
@@ -1,0 +1,7 @@
+// Game Start Settings
+export default {
+	type: Phaser.AUTO,
+	parent: "HTML5-Tower-Defense",
+	width: window.innerWidth / 1.5,
+	height: window.innerHeight / 1.5,
+};
